requirements.txt
colour==0.1.2
Pillow==3.4.2
webcolors==1.7
